SPDB_server/requirements.txt
flask-cors==3.0.10
flask-jwt-extended==4.4.4
flask-migrate==4.0.0
flask-restful==0.3.9
pip-autoremove==0.10.0
psycopg2==2.9.5
requests==2.28.1
sqlalchemy-utils==0.38.3
